[util] moving persistentcache to more suitable location.
